refactor: extract shared path validation and error utilities in ai-reviewer
Extract duplicated worktree path validation, path traversal prevention,
and error message extraction into shared utilities. All 7 tool files now
use centralized validation logic from pathValidator.ts and errorUtils.ts.

- Create pathValidator.ts with validateWorktreePath, validateRelativePath, and validatePath
- Create errorUtils.ts with errorToString
- Update all 7 tools to use shared utilities
- Eliminate ~150 lines of duplicated validation code
- Security-critical validation now centralized
