+++ packages/ai-reviewer/src/tools/readFileTool.ts
@@ -1,34 +1,12 @@
 import * as vscode from 'vscode';
 import * as path from 'path';
 import * as fs from 'fs/promises';
-import { validWorktreePaths } from './worktreeRegistry';
-
-function validateWorktreePath(worktreePath: string): string | undefined {
-  if (!validWorktreePaths.has(path.resolve(worktreePath))) {
-    return 'Invalid worktree path: not a known managed worktree';
-  }
-  return undefined;
-}
+import { validatePath } from './pathValidator';
+import { errorToString } from './errorUtils';
 
 interface ReadFileInput {
   worktreePath: string;
   filePath: string;
-}
-
-function validatePath(worktreePath: string, filePath: string): string | undefined {
-  const wtError = validateWorktreePath(worktreePath);
-  if (wtError) return wtError;
-
-  const normalized = path.normalize(filePath);
-  if (normalized === '..' || normalized.startsWith('..' + path.sep) || path.isAbsolute(normalized)) {
-    return 'Path traversal not allowed: filePath must be relative and within the worktree';
-  }
-  const resolved = path.resolve(worktreePath, normalized);
-  const root = path.resolve(worktreePath);
-  if (!resolved.startsWith(root + path.sep) && resolved !== root) {
-    return 'Path traversal not allowed: resolved path escapes the worktree';
-  }
-  return undefined;
 }
 
 export function registerReadFileTool(): vscode.Disposable {
@@ -66,7 +44,7 @@ export function registerReadFileTool(): vscode.Disposable {
           new vscode.LanguageModelTextPart(content),
         ]);
       } catch (err) {
-        const msg = err instanceof Error ? err.message : String(err);
+        const msg = errorToString(err);
         const isNotFound =
           msg.includes('ENOENT') ||
           msg.includes('FileNotFound') ||
